Moved model table metadata and compressed segments test util
