+++ crates/modelardb_common/src/test/data_generation.rs
@@ -25,9 +25,7 @@ use rand::distributions::Uniform;
 use rand::rngs::StdRng;
 use rand::{Rng, SeedableRng};
 
-use crate::types::{
-    Timestamp, TimestampArray, TimestampBuilder, ValueArray, ValueBuilder,
-};
+use crate::types::{Timestamp, TimestampArray, TimestampBuilder, ValueArray, ValueBuilder};
 
 /// Randomly generated static seed for the random number generators used for all data generation.
 /// One randomly generated seed is used to ensure new data is generated each time the tests are
