+++ crates/modelardb_server/src/query/mod.rs
@@ -554,16 +554,15 @@ mod tests {
 
     use datafusion::logical_expr::lit;
     use datafusion::prelude::Expr;
+    use modelardb_common::test;
     use modelardb_common::types::Timestamp;
-
-    use crate::common_test;
 
     const TIMESTAMP_PREDICATE_VALUE: Timestamp = 37;
 
     // Tests for rewrite_and_combine_filters().
     #[test]
     fn test_rewrite_empty_vec() {
-        let schema = common_test::model_table_metadata().schema;
+        let schema = test::model_table_metadata().schema;
         let (parquet_filter, grid_filter) = rewrite_and_combine_filters(&schema, &[]);
         assert!(parquet_filter.is_none());
         assert!(grid_filter.is_none());
@@ -572,7 +571,7 @@ mod tests {
     #[test]
     fn test_rewrite_greater_than_timestamp() {
         let filters = new_timestamp_filters(Operator::Gt);
-        let schema = common_test::model_table_metadata().schema;
+        let schema = test::model_table_metadata().schema;
         let (parquet_filter, grid_filter) = rewrite_and_combine_filters(&schema, &filters);
 
         assert_binary_expr(
@@ -593,7 +592,7 @@ mod tests {
     #[test]
     fn test_rewrite_greater_than_or_equal_timestamp() {
         let filters = new_timestamp_filters(Operator::GtEq);
-        let schema = common_test::model_table_metadata().schema;
+        let schema = test::model_table_metadata().schema;
         let (parquet_filter, grid_filter) = rewrite_and_combine_filters(&schema, &filters);
 
         assert_binary_expr(
@@ -614,7 +613,7 @@ mod tests {
     #[test]
     fn test_rewrite_less_than_timestamp() {
         let filters = new_timestamp_filters(Operator::Lt);
-        let schema = common_test::model_table_metadata().schema;
+        let schema = test::model_table_metadata().schema;
         let (parquet_filter, grid_filter) = rewrite_and_combine_filters(&schema, &filters);
 
         assert_binary_expr(
@@ -635,7 +634,7 @@ mod tests {
     #[test]
     fn test_rewrite_less_than_or_equal_timestamp() {
         let filters = new_timestamp_filters(Operator::LtEq);
-        let schema = common_test::model_table_metadata().schema;
+        let schema = test::model_table_metadata().schema;
         let (parquet_filter, grid_filter) = rewrite_and_combine_filters(&schema, &filters);
 
         assert_binary_expr(
@@ -656,7 +655,7 @@ mod tests {
     #[test]
     fn test_rewrite_equal_timestamp() {
         let filters = new_timestamp_filters(Operator::Eq);
-        let schema = common_test::model_table_metadata().schema;
+        let schema = test::model_table_metadata().schema;
         let (parquet_filter, grid_filter) = rewrite_and_combine_filters(&schema, &filters);
 
         if let Expr::BinaryExpr(BinaryExpr { left, op, right }) = parquet_filter.unwrap() {
